backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
PyMySQL==1.1.0
python-dotenv==1.0.1
cryptography==42.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
